Update to Google Analytics v4
Change the property id for gtag.js to the new GA4. UA is going to
be deprecated in a year.

Deprecation blog post:
https://blog.google/products/marketingplatform/analytics/prepare-for-future-with-google-analytics-4/

Technically it would have been possible to leave the `UA-` id in place
for another year since GA collects data for both GA3 (UA) and GA4 as
the tags are connected. However, since the sunset date has already
been announced, this switch would have to be performed soon anyway, so
might as well do it now (since I just created a GA4 property and got
the new `G-` id).

h