Change depolyment mecchanism for github pages, use docs folder
In the past, gh-pages aka Pages, built the site site from main branch.
This started causing issues like this:
https://github.com/nazmulidris/developerlife.com/actions/runs/7860254519/job/21447225849

Here's a likely explanation of what might be causing this:
https://stackoverflow.com/questions/75912209/build-error-while-trying-to-publish-my-github-pages

My understanding is that GitHub is moving away from teh default jekyll configuration
and pushing people to use a custom GitHub action instead, that does all the ruby,
gem, jekyll configuration manually using a Gemfile and the correct version of Ruby.
Also, Ruby version 3.0 came out recently which may have something to do w/ all the
changes in how "default" dependencies are handled. Moving forwards, dependencies have
to be explicity w/ Ruby 3. And this collides with the change to GitHub Actions for
buidling Pages.